move func to interactor
