+++ lib/data/models/category_item.dart
@@ -1,0 +1,9 @@
+class CategoryItem {
+  CategoryItem({
+    required this.name,
+    required this.isSelected,
+  });
+
+  String name;
+  bool isSelected;
+}
