+++ lib/ui/screens/add_sight_screen/add_sight_screen.dart
@@ -7,9 +7,10 @@ import 'package:places/ui/my_app/routes.dart';
 import 'package:places/ui/my_app/my_scroll_physics.dart';
 import 'package:places/ui/screens/add_sight_screen/widget_category_header.dart';
 import 'package:places/ui/screens/add_sight_screen/dialog_add_photo.dart';
-import 'package:places/ui/screens/add_sight_screen/add_sight_model.dart';
-import 'package:places/ui/screens/select_category_screen/select_category_model.dart';
+import 'package:places/data/interactors/select_category_model.dart';
 import 'package:places/ui/widgets/my_image_widget.dart';
+
+import 'package:places/data/interactors/place_interactor.dart';
 
 class AddSightScreen extends StatefulWidget {
   const AddSightScreen({Key? key}) : super(key: key);
@@ -34,7 +35,7 @@ class _AddSightScreenState extends State<AddSightScreen> {
   @override
   Widget build(BuildContext context) {
     //+
-    _listOfPhotos = context.watch<AddSightModel>().listOfPhotos;
+    _listOfPhotos = context.watch<PlaceInteractor>().listOfPhotos;
 
     //+
     final _currentlySelectedCategory =
@@ -391,7 +392,7 @@ class _AddSightScreenState extends State<AddSightScreen> {
   }
 
   void onTapOnSave() {
-    context.read<AddSightModel>().saveNew(
+    context.read<PlaceInteractor>().saveNew(
           name: textControllerName.text,
           lat: double.parse(textControllerLat.text),
           lon: double.parse(textControllerLon.text),
